requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2